requirements.txt
Flask==3.0.3
PyJWT==2.9.0
pymongo==4.8.0
python-dotenv==1.0.1
email-validator==2.2.0


